// apps/odonto-front/src/generated/ts/CreateAppointmentDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type CreateAppointmentDto = {
  /**
   * @type string
   */
  date: string;
  /**
   * @type number
   */
  duration: number;
  /**
   * @type number
   */
  dentistId: number;
  /**
   * @type number
   */
  patientId: number;
};