// apps/odonto-front/src/generated/ts/UpdateAppointmentDto.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export const updateAppointmentDtoStatusEnum = {
    SCHEDULED: "SCHEDULED",
    CONFIRMED: "CONFIRMED",
    CANCELLED: "CANCELLED",
    COMPLETED: "COMPLETED",
    ABSENT: "ABSENT"
} as const;

export type UpdateAppointmentDtoStatusEnumKey = (typeof updateAppointmentDtoStatusEnum)[keyof typeof updateAppointmentDtoStatusEnum];

export const updateAppointmentDtoCancelledByEnum = {
    PATIENT: "PATIENT",
    CLINIC: "CLINIC"
} as const;

export type UpdateAppointmentDtoCancelledByEnumKey = (typeof updateAppointmentDtoCancelledByEnum)[keyof typeof updateAppointmentDtoCancelledByEnum];

export type UpdateAppointmentDto = {
    /**
     * @type string | undefined
    */
    date?: string;
    /**
     * @type number | undefined
    */
    duration?: number;
    /**
     * @type number | undefined
    */
    dentistId?: number;
    /**
     * @type number | undefined
    */
    patientId?: number;
    /**
     * @type string | undefined
    */
    status?: UpdateAppointmentDtoStatusEnumKey;
    /**
     * @type string | undefined
    */
    cancelledBy?: UpdateAppointmentDtoCancelledByEnumKey;
    /**
     * @type string | undefined
    */
    cancellationReason?: string;
};